+++ robot_descriptions/loaders/mujoco.py
@@ -11,6 +11,8 @@ from typing import Optional
 
 import mujoco
 
+from .._cache import description_commit
+
 
 def load_robot_description(
     description_name: str,
@@ -28,11 +30,8 @@ def load_robot_description(
     Returns:
         Robot model for MuJoCo.
     """
-    if commit is not None:  # technical debt, see #31
-        os.environ["ROBOT_DESCRIPTION_COMMIT"] = commit
-    module = import_module(f"robot_descriptions.{description_name}")
-    if commit is not None:
-        os.environ.pop("ROBOT_DESCRIPTION_COMMIT", None)
+    with description_commit(commit):
+        module = import_module(f"robot_descriptions.{description_name}")
     if not hasattr(module, "MJCF_PATH"):
         raise ValueError(f"{description_name} is not an MJCF description")