+++ robot_descriptions/loaders/robomeshcat.py
@@ -8,12 +8,12 @@
 .. _RoboMeshCat: https://github.com/petrikvladimir/RoboMeshCat
 """
 
-import os
 from importlib import import_module  # type: ignore
 from typing import Optional
 
 import robomeshcat
 
+from .._cache import description_commit
 from .._xacro import get_urdf_path
 from .pinocchio import get_package_dirs
 
@@ -35,11 +35,8 @@ def load_robot_description(
     Returns:
         Robot model for RoboMeshCat.
     """
-    if commit is not None:  # technical debt, see #31
-        os.environ["ROBOT_DESCRIPTION_COMMIT"] = commit
-    module = import_module(f"robot_descriptions.{description_name}")
-    if commit is not None:
-        os.environ.pop("ROBOT_DESCRIPTION_COMMIT", None)
+    with description_commit(commit):
+        module = import_module(f"robot_descriptions.{description_name}")
     if not hasattr(module, "URDF_PATH") and not hasattr(module, "XACRO_PATH"):
         raise ValueError(f"{description_name} is not a URDF/Xacro description")
     urdf_path = get_urdf_path(module, xacro_args=xacro_args)